move all dash VS tests to single folder
