+++ .azure-pipelines/test-docker-sonic-vs-template.yml
@@ -136,7 +136,7 @@ jobs:
       fi
 
       all_tests=$(ls test_*.py | xargs)
-      all_tests="${all_tests} p4rt"
+      all_tests="${all_tests} p4rt dash"
 
       if [ -n '${{ parameters.run_tests_pattern }}' ]; then
         all_tests=" $(ls ${{ parameters.run_tests_pattern }} | xargs) "
